function-app/requirements.txt
azure-functions
azure-identity
azure-monitor-query
paramiko
